move utility function to pkgs.rs
